pyproject.toml
[project]
name = "autozhy-logisim-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.1",
    "patool>=4.0.4",
    "pdfplumber>=0.11.9",
    "psutil>=7.2.2",
    "pydantic>=2.12.5",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-docx>=1.2.0",
    "openai>=2.31.0",
    "google-genai>=1.72.0",
    "websockets>=12.0",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
